Refactor AlignBuilder to separate file
Refactored AlignmentBuilder to a separate file to clean up code.
Rename alignment_tests.rs to tests.rs to remove repetition in names.
